Use Table for numbered methods, fix space formatting
Thanks to the help of @AterAnimAvis, it was discovered that inherited
methods are also present in the TSRG; if class B extends class A which
has a method, if B overrides the method then two entries are present:
one for A's method, and one for B's override of A's method.

Before this commit, the collection used for numbered methods was an
ImmutableMap of Integers to SRGEntry.NumberedMethods. This meant that
the two entries above would be condensed into one entry in the map/db.

This commit changes the used collection to Guava's Tables, which allow
for 2 keys to identify 1 value. The table column key is the class, the
table row key is the SRG ID, and the value is the numbered method.

This commit also cleans up the spacing and formatting.